+++ CHANGELOG.md
@@ -96,6 +96,11 @@
 - sctbench/inspect_examples/bbuf.c: correct type for `argv`
 - sctbench/inspect_examples/boundedBuffer.c: correct type for `argv`
 - sctbench/inspect_examples/bbuf.c: exclude `malloc.h`
+- sctbench/inspect_examples/ctrace-test.c: include `unistd.h`
+- sctbench/inspect_examples/ctrace-test.c: include `string.h`
+- sctbench/inspect_examples/ctrace-test.c: correct assignment to `t->fmt[0]`
+- sctbench/inspect_examples/ctrace-test.c: declare `trc_stop_client()`
+- sctbench/inspect_examples/ctrace-test.c: declare `trc_start_client()`
 
 ## Added
 - build action for macOS